requirements.txt
numpy
scipy
scikit-learn
pillow
pyyaml
tqdm
matplotlib

torch
torchvision
torch-geometric
